+++ stackgvm/source/include/gvm_branch/bits_not_equal.hpp
@@ -3,11 +3,11 @@
 BCC(X_NE_LS) {
     gvmDebugOpcode(
         "bcc ne.i (%d), #%d, %d : %d != %d => ",
-        OPS(1),
-        S8(2),
+        (int)OPS(1),
+        (int)S8(2),
         (int)J16(3),
-        LOC(1).i,
-        S8(2)
+        (int)LOC(1).i,
+        (int)S8(2)
     );
     if (LOC(1).i != S8(2)) {
         gvmDebugJump(3);
@@ -23,11 +23,11 @@ BCC(X_NE_LS) {
 BCC(X_NE_I0S) {
     gvmDebugOpcode(
         "bcc.i ne.i (i0 + %u), #%d, %d : %d != %d => ",
-        OPU(1),
-        S8(2),
+        (unsigned)OPU(1),
+        (int)S8(2),
         (int)J16(3),
-        IX0(1).i,
-        S8(2)
+        (int)IX0(1).i,
+        (int)S8(2)
     );
     if (IX0(1).i != S8(2)) {
         gvmDebugJump(3);
@@ -43,11 +43,11 @@ BCC(X_NE_I0S) {
 BCC(X_NE_I1S) {
     gvmDebugOpcode(
         "bcc ne.i (i1 + %u), #%d, %d : %d != %d => ",
-        OPU(1),
-        S8(2),
+        (unsigned)OPU(1),
+        (int)S8(2),
         (int)J16(3),
-        IX1(1).i,
-        S8(2)
+        (int)IX1(1).i,
+        (int)S8(2)
     );
     if (IX1(1).i != S8(2)) {
         gvmDebugJump(3);
@@ -63,11 +63,11 @@ BCC(X_NE_I1S) {
 BCC(X_NE_LL) {
     gvmDebugOpcode(
         "bcc ne (%d), (%d), %d : 0x%08X != 0x%08X => ",
-        OPS(1),
-        OPS(2),
+        (int)OPS(1),
+        (int)OPS(2),
         (int)J16(3),
-        LOC(1).u,
-        LOC(2).u
+        (unsigned)LOC(1).u,
+        (unsigned)LOC(2).u
     );
     if (LOC(1).u != LOC(2).u) {
         gvmDebugJump(3);
@@ -83,11 +83,11 @@ BCC(X_NE_LL) {
 BCC(X_NE_LI0) {
     gvmDebugOpcode(
         "bcc ne (%d), (i0 + %u), %d : 0x%08X != 0x%08X => ",
-        OPS(1),
-        OPU(2),
+        (int)OPS(1),
+        (unsigned)OPU(2),
         (int)J16(3),
-        LOC(1).u,
-        IX0(2).u
+        (unsigned)LOC(1).u,
+        (unsigned)IX0(2).u
     );
     if (LOC(1).u != IX0(2).u) {
         gvmDebugJump(3);
@@ -103,11 +103,11 @@ BCC(X_NE_LI0) {
 BCC(X_NE_LI1) {
     gvmDebugOpcode(
         "bcc ne (%d), (i1 + %u), %d : 0x%08X != 0x%08X => ",
-        OPS(1),
-        OPU(2),
+        (int)OPS(1),
+        (unsigned)OPU(2),
         (int)J16(3),
-        LOC(1).u,
-        IX1(2).u
+        (unsigned)LOC(1).u,
+        (unsigned)IX1(2).u
     );
     if (LOC(1).u != IX1(2).u) {
         gvmDebugJump(3);
@@ -123,11 +123,11 @@ BCC(X_NE_LI1) {
 BCC(X_NE_II) {
     gvmDebugOpcode(
         "bcc ne (i0 + %u), (i1 + %u), %d : 0x%08X != 0x%08X => ",
-        OPU(1),
-        OPU(2),
+        (unsigned)OPU(1),
+        (unsigned)OPU(2),
         (int)J16(3),
-        IX0(1).u,
-        IX1(2).u
+        (unsigned)IX0(1).u,
+        (unsigned)IX1(2).u
     );
     if (IX0(1).u != IX1(2).u) {
         gvmDebugJump(3);
